requirements.txt
# InformaticaProjectAnalysis — Python dependencies
# Pin major.minor, allow patch updates

fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.5,<3.0
pyyaml>=6.0,<7.0
defusedxml>=0.7,<1.0
